initial tailwind and neumorphism stuff
