--- src/components/login/Login.module.scss
@@ -1,36 +1,0 @@
-@import '../../global.scss';
-
-.flex {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.logo_wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  user-select: none;
-  min-width: 15.625rem;
-}
-
-.logo_item:last-of-type {
-  padding: 0 0 ($unit * 2) 0;
-}
-
-.logo {
-  max-width: 100px;
-  width: 100%;
-  padding: 0 0 ($unit * 2) 0;
-}
-
-.login_form {
-  background: rgba(255, 255, 255, 0.13);
-  border-radius: 1rem;
-  border: 1px solid rgba(218, 218, 218, 0.3);
-  padding: 2.5rem;
-}
